Moving code into an  module
