+++ src/cofi_espresso/_machine/__init__.py
@@ -1,0 +1,2 @@
+### The `espresso.machine` module should contain most of the
+### 'magic' code that builds the package.
